abstracted validationMiddleware out
